requirements.txt
flask
flask-cors
flask-mail
mysql-connector-python
python-dotenv